chore: update alignments, elements sizing, etc
