+++ tailwind.config.js
@@ -26,6 +26,14 @@ module.exports = {
     './src/shared/components/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
+    screens: {
+      xs: '375px',
+      sm: '640px',
+      md: '768px',
+      lg: '1024px',
+      xl: '1280px',
+      '2xl': '1440px',
+    },
     extend: {
       colors: {
         'gray-1': 'hsl(0,0%,93.3%)',
